+++ src/recon_tmpl_16.rs
@@ -1197,7 +1197,7 @@ unsafe fn decode_coefs(
                         unreachable!();
                     }
                     let level = &mut levels[(x as isize * stride + y as isize) as usize..];
-                    ctx = get_lo_ctx(level, TX_CLASS_2D, &mut mag, lo_ctx_offsets, x, y, stride);
+                    ctx = get_lo_ctx(level, TX_CLASS_2D, &mut mag, lo_ctx_offsets, x, y, stride as usize);
                     if TX_CLASS_2D as libc::c_int == TX_CLASS_2D as libc::c_int {
                         y |= x;
                     }
@@ -1277,7 +1277,7 @@ unsafe fn decode_coefs(
                         lo_ctx_offsets,
                         0 as libc::c_int as libc::c_uint,
                         0 as libc::c_int as libc::c_uint,
-                        stride,
+                        stride as usize,
                     )
                 };
                 dc_tok = dav1d_msac_decode_symbol_adapt4(
@@ -1424,7 +1424,7 @@ unsafe fn decode_coefs(
                         lo_ctx_offsets_0,
                         x_0,
                         y_0,
-                        stride_0,
+                        stride_0 as usize,
                     );
                     if TX_CLASS_H as libc::c_int == TX_CLASS_2D as libc::c_int {
                         y_0 |= x_0;
@@ -1505,7 +1505,7 @@ unsafe fn decode_coefs(
                         lo_ctx_offsets_0,
                         0 as libc::c_int as libc::c_uint,
                         0 as libc::c_int as libc::c_uint,
-                        stride_0,
+                        stride_0 as usize,
                     )
                 };
                 dc_tok = dav1d_msac_decode_symbol_adapt4(
@@ -1651,7 +1651,7 @@ unsafe fn decode_coefs(
                         lo_ctx_offsets_1,
                         x_1,
                         y_1,
-                        stride_1,
+                        stride_1 as usize,
                     );
                     if TX_CLASS_V as libc::c_int == TX_CLASS_2D as libc::c_int {
                         y_1 |= x_1;
@@ -1732,7 +1732,7 @@ unsafe fn decode_coefs(
                         lo_ctx_offsets_1,
                         0 as libc::c_int as libc::c_uint,
                         0 as libc::c_int as libc::c_uint,
-                        stride_1,
+                        stride_1 as usize,
                     )
                 };
                 dc_tok = dav1d_msac_decode_symbol_adapt4(
