+++ src/recon_tmpl_8.rs
@@ -1169,7 +1169,7 @@ unsafe fn decode_coefs(
                         unreachable!();
                     }
                     let level = &mut levels[(x as isize * stride + y as isize) as usize..];
-                    ctx = get_lo_ctx(level, TX_CLASS_2D, &mut mag, lo_ctx_offsets, x, y, stride);
+                    ctx = get_lo_ctx(level, TX_CLASS_2D, &mut mag, lo_ctx_offsets, x, y, stride as usize);
                     if TX_CLASS_2D as libc::c_int == TX_CLASS_2D as libc::c_int {
                         y |= x;
                     }
@@ -1249,7 +1249,7 @@ unsafe fn decode_coefs(
                         lo_ctx_offsets,
                         0 as libc::c_int as libc::c_uint,
                         0 as libc::c_int as libc::c_uint,
-                        stride,
+                        stride as usize,
                     )
                 };
                 dc_tok = dav1d_msac_decode_symbol_adapt4(
@@ -1395,7 +1395,7 @@ unsafe fn decode_coefs(
                         lo_ctx_offsets_0,
                         x_0,
                         y_0,
-                        stride_0,
+                        stride_0 as usize,
                     );
                     if TX_CLASS_H as libc::c_int == TX_CLASS_2D as libc::c_int {
                         y_0 |= x_0;
@@ -1476,7 +1476,7 @@ unsafe fn decode_coefs(
                         lo_ctx_offsets_0,
                         0 as libc::c_int as libc::c_uint,
                         0 as libc::c_int as libc::c_uint,
-                        stride_0,
+                        stride_0 as usize,
                     )
                 };
                 dc_tok = dav1d_msac_decode_symbol_adapt4(
@@ -1622,7 +1622,7 @@ unsafe fn decode_coefs(
                         lo_ctx_offsets_1,
                         x_1,
                         y_1,
-                        stride_1,
+                        stride_1 as usize,
                     );
                     if TX_CLASS_V as libc::c_int == TX_CLASS_2D as libc::c_int {
                         y_1 |= x_1;
@@ -1703,7 +1703,7 @@ unsafe fn decode_coefs(
                         lo_ctx_offsets_1,
                         0 as libc::c_int as libc::c_uint,
                         0 as libc::c_int as libc::c_uint,
-                        stride_1,
+                        stride_1 as usize,
                     )
                 };
                 dc_tok = dav1d_msac_decode_symbol_adapt4(
